refactor: preserve client source boundaries
