+++ SSRVPN_MacOS/macos/Runner/AppDelegate.swift
@@ -29,7 +29,6 @@ private enum ApplicationTerminationLeaseState: Equatable {
   case pending(UUID)
   case committed
 }
-
 class AppDelegate: FlutterAppDelegate {
   private let instanceLease = AppInstanceLease()
   private let coreProcessOperationQueue = DispatchQueue(
@@ -125,70 +124,6 @@ class AppDelegate: FlutterAppDelegate {
 
   func performCoreProcessOperationAndWait(_ operation: () -> Void) {
     coreProcessOperationQueue.sync(execute: operation)
-  }
-
-  func startProxyGuardian(
-    statePath: String,
-    nonce: String,
-    readyTimeout: TimeInterval = 2
-  ) -> Bool {
-    guard
-      (statePath as NSString).isAbsolutePath,
-      ProxyGuardianCommand.isValidNonce(nonce),
-      let generation = AppDelegate.currentProcessGeneration(getpid()),
-      let executablePath = AppDelegate.currentExecutablePath(getpid())
-    else {
-      return false
-    }
-    let stateURL = URL(fileURLWithPath: statePath).standardizedFileURL
-    let canonicalExecutable = URL(fileURLWithPath: executablePath)
-      .standardizedFileURL.resolvingSymlinksInPath().path
-    let owner = ProxyGuardianOwnerIdentity(
-      pid: getpid(),
-      startSeconds: generation.startSeconds,
-      startMicroseconds: generation.startMicroseconds,
-      executablePath: canonicalExecutable
-    )
-    let guardianArguments = ProxyGuardianArguments(
-      stateURL: stateURL,
-      nonce: nonce,
-      owner: owner
-    )
-    guard !proxyStatePathEntryExists(at: guardianArguments.readyURL) else {
-      return false
-    }
-
-    let process = Process()
-    process.executableURL = URL(fileURLWithPath: canonicalExecutable)
-    process.arguments = ProxyGuardianCommand.arguments(
-      stateURL: stateURL,
-      nonce: nonce,
-      owner: owner
-    )
-    if let nullOutput = FileHandle(forWritingAtPath: "/dev/null") {
-      process.standardOutput = nullOutput
-      process.standardError = nullOutput
-    }
-    process.terminationHandler = { _ in }
-    do {
-      try process.run()
-    } catch {
-      return false
-    }
-
-    let deadline = Date().addingTimeInterval(readyTimeout)
-    while Date() < deadline {
-      if ProxyGuardianCommand.consumeReadyFile(
-        at: guardianArguments.readyURL,
-        nonce: nonce
-      ) {
-        return process.isRunning
-      }
-      if !process.isRunning { return false }
-      Thread.sleep(forTimeInterval: 0.05)
-    }
-    if process.isRunning { process.terminate() }
-    return false
   }
 
   @discardableResult
@@ -475,8 +410,7 @@ class AppDelegate: FlutterAppDelegate {
     }
     DistributedNotificationCenter.default().removeObserver(self)
     releaseInstanceLease()
-    // NSApplicationDelegate termination is an optional callback. Flutter's
-    // parent delegate has no implementation to chain to here.
+    // This optional callback has no parent implementation to chain to.
   }
 
   func tunRequestURLs(in support: URL) -> [URL] {
